Initial configuration for pre-commit with black
We change the usage of yapf plus a very customized group of external
plugins to only rely on black (for auto-linting) + flake8 (warning
checks) + isort (auto-sort import) + docformatter (auto-lint
docstring)

This configuration is way more easy to mantain since there is not too
many options to configure the auto linting.

The only configuration files are the ones for the warning checks that
can be adapted/tunned over time if needed.