Make `--running` flag the default for `shell`, `manage` and `test`
Now, when calling `inv docker.shell` (and others) all the commands will be
executed in the container that it's already running (using all the context
already specified, like `--webpack` or `--ext-theme`). For this to work, it's
required to have called `inv docker.up` previously, otherwise the command will
fail saying:

ERROR: No container found for web_1

If you still want to run the command anyways, you can add `--no-running`
argument to the invoke command to spin up a new container just to run that
command.