Backup PostgreSQL database when running `migrate`
When calling `inv docker.manage migrate` we can pass `--backupdb` to save an SQL
file with the content of the database before running the migration.

The name of the file uses the current date plus the git commit. This allows us
to go back to the previous state in case it's needed.