Use `CTRL-p CTRL-p` to detach an attached container
When attaching a container with:

inv docker.attach web

you can now detach it without killing the process with `CTRL-p CTRL-p`. The
default sequence is CTRL-p CTRL-q for containers running with `-i -t` but it
seems that CTRL-q does not get recognized by Invoke for some reason. So, I'm
using CTRL-p CTRL-p for this.

https://docs.docker.com/engine/reference/commandline/attach/