Use nodemon for reloading on all server processes
This examples nodemon from the celery entrypoint script to the web and
proxito service entrypoint scripts.

The Django runserver --reload option doesn't see to be configurable, and
seems to watch a lot of files that are not needed for reloading. I can't
tell what exactly is being watched by runserver, but I am getting
frequent `too many open files` errors from processes in Docker. Running with
--no-reload mostly solves this issue, but requires manually restarting
services.

Someone else should spend some time with this, I'm not confident that this will:

1. Fix the issue for others
2. Not cause irreparable damage to your local installation