Raise nofile ulimit for web/proxito containers
I'm hitting "too many open files" rather consistently (see #160 for a
first try at resolving this). I am getting some inconsistent results
still, so not sure if this solves the problem or not.

I was just able to start the development env without altering the reload
mechanism though, so perhaps this just needs to be bumped up a bit.

I'm also not opposed to changing the reload mechanism too though.