Used modern Node for webpack, and us npm ci instead
This was using a crusty old Node version and was calling `npm install`,
so the lock file was being updated and reverted to version 1. This was
causing significant spin up time for the docker container.

`npm ci` installs the exact dependencies used in the last build, so we
can ensure a CI passing build will match the prod build closely.