Development: argument to set logging level
It was a little annoying that we had the logging level set to DEBUG since we are
flooded wth a lot of messages, even when debugging things.

I set the default logging level to INFO, but added a `log-level=` argument to
easily change it in case we want to be more verbose while working locally.