Dev: Pass http-domain to RTD_PUBLIC_DOMAIN too
acc68e51 added the ability to customize RTD_PRODUCTION_DOMAIN
using the --http-domain argument to docker.up

The Docker Compose settings also allows RTD_PUBLIC_DOMAIN to be
configured using an env var but this appears to be unused. Pass
http-domain to RTD_PUBLIC_DOMAIN too so that the generated docs
can be tested on a custom domain too.

Pass RTD_PUBLIC_DOMAIN to the containers wherever
RTD_PRODUCTION_DOMAIN is used.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>