Fix issues with response return order
If `Response.clone()` is used without `await` the resulting instance has
missing attributes, like `url = null`. The docs do not indicate that
`clone()` is async however, so it's not clear why this is required.

With this, and more careful response instance usage, most of the cases
of dangling, unused, cloned responses have been removed. Now, any
response that is used is cloned right as we use it and evaluate the
body. This is important, as `response.body` is ultimately what causes
the request to buffer.