Clean up addons injection script and fix bugs in library (#227)
* Move addons injection files into package structure

We should still figure out moving these files in

- #226

* Add supplemental files to support updated addons injection module

* Update injection script

* Fix issues with response return order

If `Response.clone()` is used without `await` the resulting instance has
missing attributes, like `url = null`. The docs do not indicate that
`clone()` is async however, so it's not clear why this is required.

With this, and more careful response instance usage, most of the cases
of dangling, unused, cloned responses have been removed. Now, any
response that is used is cloned right as we use it and evaluate the
body. This is important, as `response.body` is ultimately what causes
the request to buffer.

* Update throw debug header

* Add more comments on response cloning

* Bump up conditional for debug error throw

* Test more response header passthrough

* Add test for build.commands builds

* Add path to wrangler build context

* Fix wrangler path in docker compose

* Increase logging on wrangler