CF Worker: remove unused `X-RTD-Hosting-Integrations` HTTP header
This header was used for `build.commands` when we were testing addons on those
projects that were using our own Sphinx extension to inject the flyout.

That is not required anymore since we have all the projects using the new
addons now.

Related https://github.com/readthedocs/readthedocs.org/pull/11653
Closes https://github.com/readthedocs/common/pull/223