Follow current pattern for wrangler container (#238)
Follow current pattern for `wrangler` container

I update the wrangler Docker configuration to follow the pattern we already have
for other containers. It mounts only the files that are needed to work and
nothing else.

This is done as an alterntive to https://github.com/readthedocs/common/pull/232
that should work for everybody.