Docker: build only updated dependencies
Allow developers to "re-build" the image only updating the Python (via `pip`)
and NodeJS dependencies (via `npm ci`).

After updating these dependencies, the resulting Docker image is commited, so
the chagnes are saved for the next run.

This reduces the amount of times we need to perform a full rebuild.

The option added is:

```
inv docker.build --dependencies
```

and

```
inv docker.up --npm-ci
```