Run application from `uv` (#257)
* Run application from `uv`

When running on Ubuntu 24.04, we cannot use `--system` anymore because pip
complains about it. I understand we are forced to create a virtualenv when
creating the Docker image.

Due to that, we need to run our application using that virtualenv we created at
build time. We use `uv run` for that.

* Run our tasks with `uv run`

* Remove install from common.sh