Addons: always inject META tag if the value exists
We were checking for `projectSlug` and `versionSlug` to exists before adding all
the META tags. When one of them doesn't exist the `resolverFilename` and
`httpStatus` were added.

Now, I'm checking for each of the values and adding them if they exist.

I found this issue while working on the analytics addons, since the `httpStatus`
META tag wasn't added because I was accessing an invalid URL without version.